Fix enrichment

+++ chao_fan/cli.py
@@ -1,4 +1,6 @@
-from .db import SQLModel, engine
+from sqlmodel import SQLModel
+
+from .db import engine
 
 
 def setup_db():
@@ -6,5 +8,11 @@ def setup_db():
 
 
 def reset_db():
+    check = input(
+        "This will delete all data in the database. Are you sure you want to continue? (yes/no)"
+    )
+    if check != "yes":
+        print("Aborting")
+        return
     SQLModel.metadata.drop_all(engine)
     SQLModel.metadata.create_all(engine)
